+++ src/renderer/components/animator/StatusToolbar/StatusToolbar.css
@@ -1,0 +1,3 @@
+.status-toolbar {
+  justify-content: center;
+}
