+++ src/renderer/components/animator/TitleToolbar/TitleToolbar.css
@@ -1,0 +1,3 @@
+.title-toolbar {
+  background-color: var(--ba-black);
+}
